Create Dialog for User Registration + Fix ChrLims
Created a layout for the dialog to use that has fields with proper character limits.

Username: 8
Phone: 19
Email: unlimited

Logic for validating final entry of user data has been implemented and the positive button result for the dialog is implemented now. Need to add textListeners and the negative button logic.